update python aws utils
